od : Potential fix for parsing command argumnets on windows
Summary: There is "--quiet" parameter in command argument whose handling was notpresent earlier on windows. As a result it did not parse any other arguments after this. Because of this the test program doesn't receive .par files on windows.

Test Plan: Tested with test_angle_computer in my local environment.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D1344

git-svn-id: https://svn.opendtect.org/od/trunk@30486 a985855d-d384-4b45-ae33-31a410998c3c