CMakeModule changes to support other build-systems on windows
Summary:
On windows, we have large problems with the builds, as they fail for no apparent reason randomly.
To test other build-systems, changes has to be incorporated (-isystem flag).

Also, FindosgGeo has changed, with changed variable name. This has to be adapted to.

Last, PDB output directory can be set in CMAKE, so it does not have to be set from outside.

Test Plan: These changes (seem to ) work in od.

Reviewers: nageswara

Reviewed By: nageswara

Differential Revision: https://phabricator.dgbes.com:8443/D1352

git-svn-id: https://svn.opendtect.org/od/branches/od4.6@30488 a985855d-d384-4b45-ae33-31a410998c3c