Fixed problem in detachCB
Summary: If detachCB was called for notifier that was already deleted, it lead to a crash (memory violation). Made detachCB check if notifier is still subscribed to, and only if it is, the cb will be removed.

Test Plan: Created test-program routine that reproduces the crash. Crash is gone after change. Tested for threading violations, and came out clean.

Reviewers: mahant

Reviewed By: mahant

Differential Revision: https://phabricator.dgbes.com:8443/D1348

git-svn-id: https://svn.opendtect.org/od/branches/od4.6@30489 a985855d-d384-4b45-ae33-31a410998c3c