Enabled warning that found the ibm-format bug.
Summary: After some detective's work, I found the bug that causes the ibm-format problem. Warning is enabled, and the decrease of optimization is removed. The original problem is however not solved, so you have to put someone on solving that.

Test Plan: Warnings are triggered in ibmformat.cc and other places.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D1362

git-svn-id: https://svn.opendtect.org/od/branches/od4.6@30513 a985855d-d384-4b45-ae33-31a410998c3c