Added check to see if class documentation is generated.
Summary: If the build fails, no programmers documentation will be generated. This script will still upload the empty documentation directory to the webservers, leading to an empty website. This revision adds a check to ensure that there is a docuemntation to be uploaded.

Test Plan: Not tested, can only be seen in nightly build.

Reviewers: nageswara

Reviewed By: nageswara

Differential Revision: https://phabricator.dgbes.com:8443/D1357

git-svn-id: https://svn.opendtect.org/od/branches/od4.6@30529 a985855d-d384-4b45-ae33-31a410998c3c