od: Restructured uiODViewer2D.
Summary: Made all added DataPacks to be removed before setting the new ones for changing the position. Made DataPacks to be created only for the mode that is visible while changing position. Disabled position change if the new position has no data to be displayed for both wva and vd. Fixed bug with Mantis ID 1980. Removed creation of uiSlicePos2DView in unwanted cases.

Test Plan: Works as expected.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D1246

git-svn-id: https://svn.opendtect.org/od/trunk@30574 a985855d-d384-4b45-ae33-31a410998c3c