od4.6: Unreachable code.
Summary: Unable to remove a temporary/unsaved marker if it is not the latest one that is added. Unable to remove a saved marker if there is atleast one unsaved marker in the edit markers dialog. Hence the fix.

Test Plan: Now I'm able to remove all kinds of markers.

Reviewers: bert

Reviewed By: bert

CC: raman, ranojay

Differential Revision: https://phabricator.dgbes.com:8443/D1444

git-svn-id: https://svn.opendtect.org/od/branches/od4.6@30622 a985855d-d384-4b45-ae33-31a410998c3c