od4.4: Created a function.
Summary: To remove DataPacks from multiple places. There are cases where destructor is called without closing window and where window is deleted by the subclass (D1172) before the DataPacks are released by the destructor of base class. Hence calling this function from the winCloseCB function as well. This also gets rid of the problem, DataPacks are not released by the WCP as one release statement is in destructor ( which is currently not called ) of FlatView::Viewer.

Test Plan: The DataPacks are being properly released by WCP and 2D Viewers.

Reviewers: raman

Reviewed By: raman

CC: aneesh, ranojay

Differential Revision: https://phabricator.dgbes.com:8443/D1173

git-svn-id: https://svn.opendtect.org/od/branches/od4.4@30656 a985855d-d384-4b45-ae33-31a410998c3c