od: Included a few function calls.
Summary: Calling removeAvailablePacks() from winCloseCB as well so as to remove DataPacks, before the viewwin() is deleted by the destructor of Well Correlation Panel. Included other calls as well, as they were included before and might pose some problems if they are deleted after the viewwin() is deleted.

Test Plan: The DataPacks are properly being released in all cases.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D1176

git-svn-id: https://svn.opendtect.org/od/trunk@30658 a985855d-d384-4b45-ae33-31a410998c3c