od: Attached a callback function.
Summary: To the mainviewer so as to fit to screen on resizing the window. Things will now be easier as the user will not have to drag scrollbars/click fit to screen button after resize.

Test Plan: Works as expected on resize. Tested in Well Correlation Panel and Prestack view in SynthRock.

Reviewers: satyaki, raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D1224

git-svn-id: https://svn.opendtect.org/od/trunk@30704 a985855d-d384-4b45-ae33-31a410998c3c