Temporary quieted warning about uninitialized variable
Summary: The trigger() function give warning that c is not initialized, which is true. The function is removed entirely in od, but we cannot do that in od4.6. Hence, we quiet the warning.

Test Plan: Not tested as such, result will be seen when committed in nightly build. The change builds nicely in VS, and there is no functional change.

Reviewers: ranojay

Reviewed By: ranojay

Differential Revision: https://phabricator.dgbes.com:8443/D1238

git-svn-id: https://svn.opendtect.org/od/branches/od4.6@30729 a985855d-d384-4b45-ae33-31a410998c3c