od: Low bitmap resolution.
Summary: After making this change, bitmap resolution in Well Correlation Panel is much better ( for the case there are more than 2 viewers. ) However, the resolution for wiggle display is slightly bad in some cases like the viewer is very narrow or something like that. Source of this problem could be something else. However the current changes are still needed.

Test Plan: Works as described in summary.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D1321

git-svn-id: https://svn.opendtect.org/od/trunk@30861 a985855d-d384-4b45-ae33-31a410998c3c