od: setting current item in ListBox
Summary: In Qt 453 if you check/unckeck an item in the list, it also becomes the current item and get 'selected'. In Qt 484 that is not the case. You can select an item and then check another item. The current item is still the one selected. This is quite confusing in user interfaces like the Installer and Attribute-Attribute 2D Crossplot. I would have preferred a Qt-based solution by setting some flags. But I couldn't find one. Hence this change.

Test Plan: Works as expected. tested in Installer and Attr=Attr crossplot 2D UI.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D1357

git-svn-id: https://svn.opendtect.org/od/trunk@30936 a985855d-d384-4b45-ae33-31a410998c3c