od4.6: Made 2D line selection work for Surface reader
Summary: The line subselection is not used when you load the horizon from tree. But if you load a HorizonCube on one line, there is no reason why the framework horizons should be loaded for all the lines.

Test Plan: Tested in ENOC survey by loading a HorizonCube with 200+ lines where it does make a difference in HorizonCube loading time.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D1481

git-svn-id: https://svn.opendtect.org/od/branches/od4.6@30984 a985855d-d384-4b45-ae33-31a410998c3c