od: Fix for error in restore session when session file includes deleted attribute data from he survey.
Summary:
Fix for bug id 2067.
While saving session it should not save those attributes which are not saved on the disk or basically which are deleted by the user after adding it to the scene.

Test Plan: Display a attribute on a inline and then display another attr. then delete the first attr. and save the session. After this fix first attr. will not be save d in file.

Reviewers: raman

Reviewed By: raman

CC: helene

Differential Revision: https://phabricator.dgbes.com:8443/D1524

git-svn-id: https://svn.opendtect.org/od/trunk@31040 a985855d-d384-4b45-ae33-31a410998c3c