od : Fix for mantis bug id 1734 ( Texture from Texture causes OD to crash )
Summary:
This is rather general and is not limited to texture attribute.

In case a user selects a similarity attribute( say sim_in ) and then tries to create another similarity attribute ( say sim_out ) whose input is previous similarity attribute 'sim_in' and in the process clicks on "ok" button instead of "Add as new", an error message pops up saying the "Input data is dependent on itself". Then if the user selects "Add as new" then 'sim_out' is added to the attribute window along with 'sim_in'.

Now when "sim in" is added on an inline, an error message pops up saying "Input is dependent on itself for sim_in attribute". But when "sim_out" is added on an inline, od crashes.

This happens as newprov is equal to inputprovider. So Provider::updateInputReqs calls Provider::setDesBufStepout which again calls Provider::updateInputReqs. This goes into an infinite loop and od crashes saying "stack overflow".

Test Plan: Tested by adding similarity and texture attribute with the specified steps. No more crash and the given error message pops up.

Reviewers: helene

Reviewed By: helene

Differential Revision: https://phabricator.dgbes.com:8443/D1528

git-svn-id: https://svn.opendtect.org/od/trunk@31058 a985855d-d384-4b45-ae33-31a410998c3c