Bugfix: Several types of input commands were not recorded in case an editing change by the user made its notifier trigger the popup of a window.
Summary:
Several types of input commands were not recorded in case an editing change by  
the user made its notifier trigger the popup of a window. The editing finished  
notification was ignored by the command composer, because it occurred nested    
within in the editing change notification. Example: Inline->Add, increase the   
slice position spinbox, and the seismic selection dialog pops up. The click OK  
was recorded, but the spinbox increase was not. Therefore, running the script   
was not successful.

Test Plan:
Recording and running the spinbox example above performs fine after the fix.    
Also identified which other commands could suffer from the same problem:        
lineedit and slider commandxs. Tested this by temporarily adding a dummy        
message pop up in arbitrary OpendTect callback functions that are triggered     
by either a change in the input field of a slider move. Recording and running   
ok after the fix. Added dummy code removed afrterwards.

Reviewers: helene

Reviewed By: helene

Differential Revision: https://phabricator.dgbes.com:8443/D1568

git-svn-id: https://svn.opendtect.org/od/branches/od4.6@31076 a985855d-d384-4b45-ae33-31a410998c3c