od: Empty log panels at the end after session restore.
Summary: Mantis ID: 1923. Extra Panels are added at the end in WCP after session restore if the current session has more than one panel for any of its wells. This is because usePar is called twice, first for reading the settings of current session and next for reading the settings of restored session.

Test Plan: Now empty panels are not added at the end after restoring a session.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D1567

git-svn-id: https://svn.opendtect.org/od/trunk@31084 a985855d-d384-4b45-ae33-31a410998c3c