More reliable method to find software directory automatically (od4.6)
Summary:
Since od4.6, we (should) no longer be dependent on the DTECT_DIR environemnt. Instead, we should always be able to compute the correct location based on the location of the executable.

With the installer, the directory structure is somewhat different, and the -5 that was used does not work. Hence, we start at the executable, and move upwards until we find the directory that contains "relinfo".

Test Plan: Tested on mac. Every executable will fail if this one fails so we will know soon if there is a problem. The old code worked fine on lux and win (but not on the instaler). The relinfo is present in both development and release environment, so that is expected to work.

Reviewers: bert

Reviewed By: bert

Differential Revision: https://phabricator.dgbes.com:8443/D1260

git-svn-id: https://svn.opendtect.org/od/branches/od4.6@31088 a985855d-d384-4b45-ae33-31a410998c3c