Several improvements to uiTreeViewItem::translate() functionality in od/.
Summary:
(1) Made GoogleTranslate and CmdDriver's name-tooltip no longer interfere.
(2) Trigger retranslation if tooltips are changed after translation.
(3) Implemented translation for subitems and multiple columns as well.

Test Plan:
Tested mixed use of GoogleTranslate and CmdDriver's name-tooltip functionality.
Tested whether translations refresh also when tooltip is changing. Analogous
modifications concerning (1) and (2) were also made for uiObject and uiAction
but that was done in multiple revisions. Review if you like, but it is mostly
to show you what kind of modifications have been made to the translate()
functionality. No urgency to port all this to the stable.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D1558

git-svn-id: https://svn.opendtect.org/od/trunk@31100 a985855d-d384-4b45-ae33-31a410998c3c