od: Crash while accessing a deleted object.
Summary: setupgrp_ should be made zero in trackerWinClosedCB function so as to avoid accessing it from the destructor after it is deleted. Hence the change.

Test Plan: Started tracking on a 2D Viewer, closed the 2D Viewer without closing the Tracking Setup dialog. Changed the survey after that ( Destructor of uiMPEPartServer will be called ). There is no crash at this point after the change.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D1611

git-svn-id: https://svn.opendtect.org/od/trunk@31109 a985855d-d384-4b45-ae33-31a410998c3c