od: Using /Zi(Program database)flag to store extended debug information in the .pdb file.
Summary: This helps to better understand the crash report, not only just call-stack but even point to the .cc file with line number information.

Test Plan: Tested with and without this flag, and found that the debugger could actually identify crash loaction in the source code from the mini dump file(.dmp) when used with this flag.

Reviewers: kristofer, aneesh

Reviewed By: aneesh

CC: aneesh

Differential Revision: https://phabricator.dgbes.com:8443/D1605

git-svn-id: https://svn.opendtect.org/od/trunk@31135 a985855d-d384-4b45-ae33-31a410998c3c