od : Fix for mantis bug id 2106 ( Crash when adding default data for inline on F3 Demo with incorrect Z range )
Summary:
The Z range is not correct as described in the note added by Helene.
http://bugs.dgbes.com/view.php?id=2106

The Z range ( Z range: 0`1848`4`D ) which is taken as a depth is considered in meters.
So here desiredvolume_ range is ( 0-1848 with step 4 ) and storedvolume_ range is( 0-1.848 with step 0.004 ).
And it creates array whose zrange is 0-1848 with 0.004 steps ( 462001 samples )...and it creates lot of havoc.

So added a check which gives user an appropriate warning and does not try to create the data cube with 462001 samples.

Test Plan:
Tested with F3_Demo, SEG_Salt2 and Penobscot by adding different types of time, depth, wheeler scene and stepping through the code to make sure zstepratio and zstepwrong are computed correctly.

Also tested with the incorrect .survey file to make sure user gets the appropriate error message.

Reviewers: helene, raman

Reviewed By: helene

Differential Revision: https://phabricator.dgbes.com:8443/D1801

git-svn-id: https://svn.opendtect.org/od/trunk@31322 a985855d-d384-4b45-ae33-31a410998c3c