od4.6: Fixed crash in evaluate attribute
Summary: If you calculate attribute slices on another element then change the selection in tree to a horizon display and move the slider in 'Evaluate attribute' dialog' you get a crash as it tries to set a texture that does not exist. Just added a check while setting texture now.

Test Plan: No more crash after the fix.

Reviewers: jaap

Reviewed By: jaap

Differential Revision: https://phabricator.dgbes.com:8443/D1843

git-svn-id: https://svn.opendtect.org/od/branches/od4.6@31411 a985855d-d384-4b45-ae33-31a410998c3c