od4.6: Crash while loading a model in SynthRock.
Summary: Due to many setPack and removePack functions being called while loading models, sometimes there are tasks that have started but before their execution is completed, the packs are getting deleted. So added a check to return if the Array2D that's being used by the task does not match with the new/current pack that's whose data is to be displayed in FlatViewer, gen_->fill() is not called.

Test Plan: Works well. Tested by loading the dataset/model which is causing the crash many times. Tested other cases as well. Everything seems to work well.

Reviewers: satyaki, raman

Reviewed By: satyaki

Differential Revision: https://phabricator.dgbes.com:8443/D1899

git-svn-id: https://svn.opendtect.org/od/branches/od4.6@31453 a985855d-d384-4b45-ae33-31a410998c3c