od : Fix for incomplete call stack with linux crash reports.
Summary:
Some of the crash reports on linux were not complete. The reason is presence of '&' in call stack/crash report. This '&' sign is used to separate different set of key-value pairs in application/x-www-form-urlencoded. So if '&' is present in the call-stack everything after this does not come with crash report.

I don't understand everything about all the different mime types. But this change seems to fix this problem.

Test Plan:
Tested this on windows and linux.
On windows called this function _post instead of _postfile and sending the mail.
Also tested on linux by calling od_uiReportIssue and sending mail locally.

Reviewers: kristofer, raman

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D1902

git-svn-id: https://svn.opendtect.org/od/trunk@31455 a985855d-d384-4b45-ae33-31a410998c3c