od: Bug fix in well track import from ascii.
Summary:
If one of track or MD is not defined in a file then well is not imported properly. In that case (if one of them is defined)track and MD will be equal.
Bug fix for id 2163

Test Plan: tested by importing well in which track is not defined.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D2081

git-svn-id: https://svn.opendtect.org/od/trunk@31666 a985855d-d384-4b45-ae33-31a410998c3c