od: Implemented New survey dialog window
Summary:
Add a dialog that is interleaved between uiSurvey and uiSurveyInfoEditor,
with the essential parameters to be filled in. I should probably also add a ZDomain and Zunit selection for the completness (not yet included). Upon exit the selected survey info provider is called. Could not be done by this new dialog, since it needs to uiSurveyInfoEditor dialog to work.

Test Plan: tested for 2D, 3D, manual selection

Reviewers: nanne, bert

Reviewed By: bert

CC: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D1635

git-svn-id: https://svn.opendtect.org/od/trunk@31765 a985855d-d384-4b45-ae33-31a410998c3c