Adding virtual destructors.
Summary: We have a new warning on the nightly builds on mac that tells us when we are missing virtual destructors on classes. Fixed the warnings.

Test Plan:
No more warnings. I cannot see possible downside. We may get crashes now as we are calling destructors that were previously not called, but then it is because of crap elsewhere.

Also fixed lint-errors (too long lines).

Reviewers: bert

Reviewed By: bert

Differential Revision: https://phabricator.dgbes.com:8443/D2143

git-svn-id: https://svn.opendtect.org/od/trunk@31785 a985855d-d384-4b45-ae33-31a410998c3c