[od4.6] Traces are resampled on the basis of last zvalue of Layer Model
Summary: Earlier was taking the last time of the largest D2TModel. But due to a function mergeSameLayers() in ElasticModel all the similar layers were clubbed into single layer. as a result the last recorded reflecltivty was higher & the lasttime in d2tmodels were the last reflectivity. But what we want is the last time of the last zvalue of the layermodel

Test Plan: Now there is a sync between both the viewers in some models in profile models( which had these problem) in F3_Demo

Reviewers: bert

Reviewed By: bert

Differential Revision: https://phabricator.dgbes.com:8443/D2221

git-svn-id: https://svn.opendtect.org/od/branches/od4.6@31900 a985855d-d384-4b45-ae33-31a410998c3c