od4.6: Restore ABI compatibility in uiTable
Summary:
although D2050 was meant to be ABI compatible according to the reviewer (Bert),
in the new release 4.6.0b1 loading an MPSI attribute in the set generates a crash in uiTable. This
is an attempt to solve this crash

Test Plan: functionality still works. Will need to make a release with it to test the MPSI plugin with this base code

Reviewers: raman

Reviewed By: raman

CC: helene

Differential Revision: https://phabricator.dgbes.com:8443/D2220

git-svn-id: https://svn.opendtect.org/od/branches/od4.6@31903 a985855d-d384-4b45-ae33-31a410998c3c