Created atomicSetIfEqual for simple setting of volatile integers and made functions for simple spinlocks.
Summary:
There is a need to create extremely simple spinlocks from integers, and then a
non-class function to do the job is needed. Made the class-based atomics
use this funciton as well.

Test Plan: The atomics continue to work, so the function is still in operation. Also added new test-case to ensure the new lock functions works.

Reviewers: bert

Reviewed By: bert

Differential Revision: https://phabricator.dgbes.com:8443/D2230

git-svn-id: https://svn.opendtect.org/od/trunk@31920 a985855d-d384-4b45-ae33-31a410998c3c