Made zdomain.cc threadsafe.
Summary:
I would normally not give this to you Bert (I've been sending you lots of stuff lately), but as we discussed the setIfEqual stuff and similar, I wanted you to see this as it is an application of it.

In this case, we call the PtrMan<>::setIfNull() function, which internally calls setIfEqual(p,0);

Test Plan: Tested, and normal operations work fine. Was not able to produce a race-condition to test that.

Reviewers: bert

Reviewed By: bert

Differential Revision: https://phabricator.dgbes.com:8443/D2257

git-svn-id: https://svn.opendtect.org/od/trunk@31992 a985855d-d384-4b45-ae33-31a410998c3c