od4.6: Fixed crash in Fault display
Summary: I am not sure if this is the best fix. The problem is the order of function calls here. If we remove the polygonFinishedCB callback first and then call showManipulator( false ), it ends up calling setStickSelectMode() and the callback is added back even though the object gets deleted. This leads to an obvious crash on Windows. So, now I just changed the order which makes sure the callback is really removed.

Test Plan: Checked in the debugger that the callback of the deleted object is no longer with the notifier.

Reviewers: jaap

Reviewed By: jaap

Differential Revision: https://phabricator.dgbes.com:8443/D2245

git-svn-id: https://svn.opendtect.org/od/branches/od4.6@32000 a985855d-d384-4b45-ae33-31a410998c3c