Added isClockWise() function to InlCrlSystem class.
Summary:
On behalf of /od_osg, but to be committed in /od to avoid merging problems!
(1) Moved isClockWise() function from SurvInfo to InlCrlSystem class.           
(2) Moved the implementation of a couple of other InlCrlSystem functions to     
the appropriate source file. Also removed the implementation of SurvInfo        
functions with the same name, and mapped these onto the InlCrlSystem functions  
to eliminate duplicated code.

Test Plan:
1) Checked isClockWise() result for several surveys, among which F3_Demo and   
Tutorial, which are known to have different senses. Also changed the coordinate 
settings in the survey setup of F3_Demo temporarily to have its sense flip.     
(2) Experienced nothing strange related to survey info when using OD.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D2275

git-svn-id: https://svn.opendtect.org/od/trunk@32034 a985855d-d384-4b45-ae33-31a410998c3c