od : Handling of m/ft conversion in case of pre-stack attribute
Summary:
https://phabricator.dgbes.com:8443/T83. This is part of task T83.
This one is bit tricky. I am not 100 percent sure about the change in cbvsreader.cc whether this is the correct way to handle this.

Expecting reviewers comments if there is a better way.

Test Plan: Stepped through the code to make sure the values we get are what we expect.

Reviewers: kristofer, raman, arnaud

Reviewed By: raman

CC: helene

Maniphest Tasks: T83

Differential Revision: https://phabricator.dgbes.com:8443/D2316

git-svn-id: https://svn.opendtect.org/od/trunk@32125 a985855d-d384-4b45-ae33-31a410998c3c