od : Fix for Crash Report ( Mantis ID 2229 : Crash in ValueSeriesGetAll::doWork() function )
Summary:
http://bugs.dgbes.com/view.php?id=2229
As per the call stack, the most likely reason for crash is when dataset_->getData() is null. Added check for the same.

Test Plan: Decreased the mChunkSize in valseries.h and was able to reproduce the crash in  RWE Dea Norway while displaying 2D Seismics. No more crash after the change.

Reviewers: raman

Reviewed By: raman

CC: helene

Differential Revision: https://phabricator.dgbes.com:8443/D2322

git-svn-id: https://svn.opendtect.org/od/trunk@32132 a985855d-d384-4b45-ae33-31a410998c3c