od : Avoiding error message in case of Voxel Connectivty Filter
Summary:
The previous default value of zero for minimum body size was not accepted as it generated error message "Minimum size is not set or is less than 1" as described by Arnaud in http://bugs.dgbes.com/view.php?id=2215.

So changing the default body size to 1 instead of zero to avoid this error message.

Test Plan: No more error message after this change.

Reviewers: arnaud, raman

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D2326

git-svn-id: https://svn.opendtect.org/od/trunk@32137 a985855d-d384-4b45-ae33-31a410998c3c