od4.6: Fix for unpacking zip archives made on Windows
Summary: While unpacking a zip archive made on Windows we do not get any file permissions bits. In that case we should simply skip the step that sets the file permissions and let the files live with the default permissions assigned by the Shell when they were created.

Test Plan: Now I can unzip a file zipped on Windows and my output is same as that generated by the unzip program available on Linux.

Reviewers: nageswara, salil

Reviewed By: salil

CC: salil

Differential Revision: https://phabricator.dgbes.com:8443/D2367

git-svn-id: https://svn.opendtect.org/od/branches/od4.6@32190 a985855d-d384-4b45-ae33-31a410998c3c