Made the scene read the old key for the z-stretch.
Summary: The "Z Scale" was still used in F3 demo, and I thought it was a good idea to still read it. Also, od4.6 had a /2 division when setting the scale & factor. As it is a number without any unit, I thought we shoudl continue to do so just to give the same end-result as we had in od4.6

Test Plan: Tested, both with new and old setting.

Reviewers: ding

Reviewed By: ding

Differential Revision: https://phabricator.dgbes.com:8443/D2430

git-svn-id: https://svn.opendtect.org/od/trunk@32299 a985855d-d384-4b45-ae33-31a410998c3c