Added version info to IOPar.
Summary:
Quite frankly, I'm not sure if this is good or not. I made this change as I had an
IOPar retrieved from a file, and my interpretation of it depended on od version. At 
the point of interpretation, the reader was long gone.

In the end, I implemented my change in a different manner. Now I'm wondering if this should be abandoned as well, or if it actually is a good idea.

Test Plan: Was tested and worked well.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D2429

git-svn-id: https://svn.opendtect.org/od/trunk@32343 a985855d-d384-4b45-ae33-31a410998c3c