Restored Coin3D defaults for manipulation behaviour. (porting of D2291 from trunk to trunk/code)
Summary:
Middle mouse button for perpendicular-to-plane translation in OSG is replaced
by left mouse button. Also restored functionality to specify the mod key masks
for in-plane and perpendicular-to-plane translation.

Test Plan:
The inl-crl-z draggers now show identical manipulation behavior as in Coin3D.
This behaviour is correctly adjusted if the default mod key masks for in-plane
and perpendicular-to-plane translation are modified via the properties
dialog or the advanced settings.


git-svn-id: https://svn.opendtect.org/od/trunk@32362 a985855d-d384-4b45-ae33-31a410998c3c