Made use of osgGeo in external directory unless osgGeo is specified manually.
Summary: If the user does not specify an osgGeo directory, it will automatically use the one checked out in the external directory and integrate that as a build.

Test Plan:
Tested on windows with visual studio.
Tested on windows with ninja
Tested on linux with Unix Makefiles
Tested on linux with Ninja
Tested on mac with XCode.

Reviewers: nageswara

Reviewed By: nageswara

Differential Revision: https://phabricator.dgbes.com:8443/D2507

git-svn-id: https://svn.opendtect.org/od/trunk@32439 a985855d-d384-4b45-ae33-31a410998c3c