[od] Fix for mantis id 2202. LAS import: meter vs feet handling is wrong
Summary:
Fix for mantis id 2202.
http://bugs.dgbes.com/view.php?id=2202

According to the bug description 
Importing a LAS file that is in FT in a survey that has depths in meters has a UI problem.

Test Plan: Tested by importing a LAS file which is in feet in a survey that has depths in metres.

Reviewers: raman

Reviewed By: raman

CC: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D2531

git-svn-id: https://svn.opendtect.org/od/trunk@32466 a985855d-d384-4b45-ae33-31a410998c3c