od4.6: Using mDefEps
Summary: Since SamplingData is used to store a variety of quantities, we cannot be sure about the range of values here. Therefore it is safer to use mDefEps instead of 1e-3. For example, the horizon writer uses SamplingData with very small steps.

Test Plan: I am able to save horizons after this fix.

Reviewers: helene

Reviewed By: helene

Differential Revision: https://phabricator.dgbes.com:8443/D2548

git-svn-id: https://svn.opendtect.org/od/branches/od4.6@32486 a985855d-d384-4b45-ae33-31a410998c3c