od4.6: Start of fix for 2D line import with decreasing trace numbers
Summary:
The CBVS system can now handle 2D lines with decreasing trace numbers. The lines end up correctly in the 2D geometry database. Alas, visualization is still not as expected.
While testing, found a problem with Simple File import that I already fixed when porting to the od_streams - I thought this one was specific to the od_streams but apparently it was not.

Test Plan: Importing simple file seismics works as expected. Also 'normal' seismics still read as expected.

Reviewers: satyaki

Reviewed By: satyaki

Differential Revision: https://phabricator.dgbes.com:8443/D2542

git-svn-id: https://svn.opendtect.org/od/branches/od4.6@32494 a985855d-d384-4b45-ae33-31a410998c3c