[od] Fixing crash when surveyinfo is not read properly
Summary:
While searching for the potential crash for mantis id 2330 found this one
http://bugs.dgbes.com/view.php?id=2330. 

When SurveyInfo::read() returns null, newsi is null. The user gets a warning but setCurrentSurvInfo( newsi ) was still called. The previous SI was deleted and cursurvinfo_ was set with this pointer, resulting in crashes at different places.

Test Plan: No more crash

Reviewers: raman

Reviewed By: raman

CC: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D2566

git-svn-id: https://svn.opendtect.org/od/trunk@32512 a985855d-d384-4b45-ae33-31a410998c3c