Added preDraw and postDraw callbacks.
Summary: These notifiers will be triggered before the scene is traversed, and after the scene is traversed. They will be used to make animations, and add a place where scene updates can be applied in a safe manner.

Test Plan: Tested as far as the notifiers are triggered at the right moment, and everything deletes nicely at closure.

Reviewers: jaap

Reviewed By: jaap

Differential Revision: https://phabricator.dgbes.com:8443/D2573

git-svn-id: https://svn.opendtect.org/od/trunk@32546 a985855d-d384-4b45-ae33-31a410998c3c