od4.6: crash fix in uiodapplmgr. Object management of data pack corrected.
Summary:
Bug fix for ID 2309.
It's a data specific crash and the data is really bad.
Data object manager didnot obtain the data pack object so it should not release the earlier object

Test Plan: tested 3D horizon tracking with the data provided for this crash and also with F3.

Reviewers: nanne

Reviewed By: nanne

CC: raman

Differential Revision: https://phabricator.dgbes.com:8443/D2575

git-svn-id: https://svn.opendtect.org/od/branches/od4.6@32602 a985855d-d384-4b45-ae33-31a410998c3c