Bugfix: HUD events like wheels must be handled before scene events
Summary:
If a HUD event callback is added to the camera, it will be handled after
the scene event callbacks (see (Composite)Viewer::eventTraversal() function).
This is why a selected scene object (e.g. an inline) was deselected after
rotating the scene with the thumb wheels. So far, I see no other option but
also no harm in adding the HUD event callback to the scene instead. Assuming
they are created in the initialization stage, they end up in the list before
the callbacks from the scene. Now the HUD event handler can put setHandled()
to true before the scene will receive these events.

Test Plan:
A selected scene object (e.g. an inline) is still deselected when clicking
the scene brackground, but is no longer deselected after rotating the scene
with the thumb wheels.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D2622

git-svn-id: https://svn.opendtect.org/od/trunk@32644 a985855d-d384-4b45-ae33-31a410998c3c