Made use of streams in cbvs read/write/management
Summary:
One question to reviewer: When writing is finished, is the best practice to do an:

return strm.isOK() as a success check, or should one do an
return !strm.isBad().

On one hand, the streams should be re-tested for soft errors, and though !isOK may have soft errors in, it's probably hard since it has been re-tried.

Test Plan: Reading is tested, writing is not tested.

Reviewers: bert

Reviewed By: bert

Maniphest Tasks: T154

Differential Revision: https://phabricator.dgbes.com:8443/D2659

git-svn-id: https://svn.opendtect.org/od/trunk@32683 a985855d-d384-4b45-ae33-31a410998c3c