od4.6: Safety against infinite recursion
Summary: This is not the ideal fix for the problem. But since the function handles only OD data files, I see no reason for a subdirectory being a symbolic link.

Test Plan: tested by creating a link inside a HorizonCube data folder that points to intself. No more recursive calls pointing to self.

Reviewers: bert

Reviewed By: bert

Differential Revision: https://phabricator.dgbes.com:8443/D2676

git-svn-id: https://svn.opendtect.org/od/branches/od4.6@32711 a985855d-d384-4b45-ae33-31a410998c3c