Green tabs got hide option and adjustable size.
Summary:
Started using osgGeo::TabBoxDragger to adjust the size of the green tabs,
and added functionality to hide them temporarily.

Test Plan:
Green tabs of box dragger now show the same size as DepthTabPlaneDragger
and that size is comparable to Coin. The tabs temporarily disappear when
decreasing the size of an already existing volume to offer the user full
view on the seismics when manouevring through the cube and indicate to
the user that the texture data does not need to be recreated or reloaded.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D2669

git-svn-id: https://svn.opendtect.org/od/trunk@32722 a985855d-d384-4b45-ae33-31a410998c3c