Differentiate between texture volume stored and ROI to be shown.
Summary:
(1) Added functionality to differentiate between the texture volume stored
and the region of interest (ROI) to be shown.
(2) Added functionality to select the texture filter.
(3) Moved away custom shader functionality to osgGeo/VolumeTechniques

Test Plan:
(1) The dimensions of a texture volume on screen can be reduced and shown
correctly without the texture being recreated or reloaded.
(2) Switched from interpolation to classification in display properties texture
tab, and indeed got the individual texture voxels on screen. The real goal was
to test whether the texture coordinates were set right, and they turned out to
be half a pixel wrong. This is fixed in visvolumedisplay.cc.
(3) Tested by intersecting volume with inline, which shows that OSG's default
fragment shader is still replaced by our custom one.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D2674

git-svn-id: https://svn.opendtect.org/od/trunk@32723 a985855d-d384-4b45-ae33-31a410998c3c