Control for adding volumes to the screen tree.
Summary:
(1) Prevent more than one volume per scene for the non-shader case,
since currently OSG cannot handle this.
(2) Insert new volume in scene tree on top of previous ones instead of
below. Cannot think of a reason why this behaviour should be different
than for sections.

Test Plan:
A second volume added to the scene is now inserted in the scene tree
above the first one in the shader case, while an info message is produced
in the non-shader case. Also tested with multiple scenes on screen.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D2661

git-svn-id: https://svn.opendtect.org/od/trunk@32726 a985855d-d384-4b45-ae33-31a410998c3c