od: Exception handling in VectorAccess
Summary:
Catching the exception and calling forceCrash. The exception out_of_range was not handled and eventually Qt did not like it. Besides, it was very difficult to debug it in case of a crash. This is what you would get in gdb:
(PE) Qt has caught an exception thrown from an event handler. Throwing

exceptions from an event handler is not supported in Qt. You must

reimplement QApplication::notify() and catch all exceptions there.



terminate called after throwing an instance of 'std::out_of_range'

  what():  vector::_M_range_check

Test Plan: In debug mode forceCrash is called while in Release mode there are no crashes as usual.

Reviewers: kristofer

Reviewed By: kristofer

CC: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D2777

git-svn-id: https://svn.opendtect.org/od/trunk@32798 a985855d-d384-4b45-ae33-31a410998c3c